Migrate vector_schema structs to dc_bundle
This commit also fixes a bunch of Rust compile warnings.
