+++ crates/figma_import/src/tools/fetch.rs
@@ -20,6 +20,7 @@ use clap::Parser;
 use dc_bundle::legacy_definition::element::node::NodeQuery;
 
 #[derive(Debug)]
+#[allow(dead_code)]
 pub struct ConvertError(String);
 impl From<crate::Error> for ConvertError {
     fn from(e: crate::Error) -> Self {
